Introduce allowUnknownMarkup config option. (#159)
Introduce `allowUnknownMarkup`, which controls whether element or attribute names unknown to HTML are allow-able.

The intent is to prevent inadvertent use of confusable markup, e.g. mixed-case script elements in a foreign namespace (`...createElementNS("bla", "sCrIpT")`). If later actions by a program were to re-parse parts of the DOM, these might be re-parsed into known HTML elements (`<script>`, in this case). This becomes urgent as a consequence of dropping case normalization.