Update the spec to match the current group consensus. (#208)
Rewrite of the API Summary, Framework, and Algorithms section, to match the current group consensus and the current explainer. More work needs to be done, but with this the spec should roughly reflect the current discussion.

This closes out a whole lot of issues: Closes #200. Closes #199. Closes #188. Closes #186. Closes #183. Closes #181. Closes #148. Closes #147. Closes #101. Closes #59. Closes #45.