Update .gitignore
Try fixing deploy issue to gh-pages.

https://github.com/WICG/sanitizer-api/actions/runs/12810876107/job/35718920474

The error appears to be that the `bikeshed` built+deploy
script performs a `git checkout`, and that won't overwrite
the just-generated files in builtins. According to git manual,
git checkout defauls to overwriting 'ignored' files. So maybe
the issue is that the 'gh-pages' branch doesn't yet know to
ignore these two files.