Handle replaceWithChildrenElements before the elements allowlist in sanitize core
Fixes #261

Currently, using a replaceWithChildrenElements list together with an elements list does not really work.
Because we first remove all elements that are not in the elements list, which means there is nothing to replace,
because it's impossible for an element to be in both lists.

Example:
```
setHTML("<p><i>hi<i></p>") { elements: ["i"], replaceWithChildrenElements: ["p"] })
```

Before this would just remove the `<p>` element. After this change they are replaced with their allowed child elements
(i.e. `<i>`).

Depends on #275